Make client a service to fix race
