+++ skymp5-client/src/index.ts
@@ -98,69 +98,6 @@ on("update", () => updateWc());
 
 once("update", verifyLoadOrder);
 
-const startClient = (): void => {
-  NetInfo.start();
-  animDebugSystem.init(settings["skymp5-client"]["animDebug"] as animDebugSystem.AnimDebugSettings);
-
-  playerCombatSystem.start();
-  once("update", () => authSystem.setPlayerAuthMode(false));
-  connectWhenICallAndNotWhenIImport();
-  new SkympClient();
-
-  once("update", verifyVersion);
-
-  let lastTimeUpd = 0;
-  const hoursOffset = -3;
-  const hoursOffsetMs = hoursOffset * 60 * 60 * 1000;
-  on("update", () => {
-    if (Date.now() - lastTimeUpd <= 2000) return;
-    lastTimeUpd = Date.now();
-
-    const gameHourId = 0x38;
-    const gameMonthId = 0x36;
-    const gameDayId = 0x37;
-    const gameYearId = 0x35;
-    const timeScaleId = 0x3a;
-
-    const gameHour = GlobalVariable.from(Game.getFormEx(gameHourId));
-    const gameDay = GlobalVariable.from(Game.getFormEx(gameDayId));
-    const gameMonth = GlobalVariable.from(Game.getFormEx(gameMonthId));
-    const gameYear = GlobalVariable.from(Game.getFormEx(gameYearId));
-    const timeScale = GlobalVariable.from(Game.getFormEx(timeScaleId));
-    if (!gameHour || !gameDay || !gameMonth || !gameYear || !timeScale) {
-      return;
-    }
-
-    const d = new Date(Date.now() + hoursOffsetMs);
-
-    let newGameHourValue = 0;
-    newGameHourValue += d.getUTCHours();
-    newGameHourValue += d.getUTCMinutes() / 60;
-    newGameHourValue += d.getUTCSeconds() / 60 / 60;
-    newGameHourValue += d.getUTCMilliseconds() / 60 / 60 / 1000;
-
-    const diff = Math.abs(gameHour.getValue() - newGameHourValue);
-
-    if (diff >= 1 / 60) {
-      gameHour.setValue(newGameHourValue);
-      gameDay.setValue(d.getUTCDate());
-      gameMonth.setValue(d.getUTCMonth());
-      gameYear.setValue(d.getUTCFullYear() - 2020 + 199);
-    }
-
-    timeScale.setValue(gameHour.getValue() > newGameHourValue ? 0.6 : 1.2);
-  });
-
-  let riftenUnlocked = false;
-  on("update", () => {
-    if (riftenUnlocked) return;
-    const refr = ObjectReference.from(Game.getFormEx(0x42284));
-    if (!refr) return;
-    refr.lock(false, false);
-    riftenUnlocked = true;
-  });
-}
-
 skillSystem.skillMenuInit();
 
 const main = () => {
@@ -172,7 +109,8 @@ const main = () => {
       new LoadGameService(sp, controller),
       new SinglePlayerService(sp, controller),
       new EnforceLimitationsService(sp, controller),
-      new SendInputsService(sp, controller)
+      new SendInputsService(sp, controller),
+      new SkympClient(sp, controller),
     ]);
   }
   catch (e) {
@@ -181,19 +119,3 @@ const main = () => {
   }
 };
 main();
-
-const authGameData = storage[AuthGameData.storageKey] as AuthGameData | undefined;
-if (!(authGameData?.local || authGameData?.remote)) {
-  authSystem.addAuthListener((data) => {
-    if (data.remote) {
-      browser.setAuthData(data.remote);
-    }
-    storage[AuthGameData.storageKey] = data;
-    spBrowser.setFocused(false);
-    startClient();
-  });
-
-  authSystem.main(settings["skymp5-client"]["lobbyLocation"] as Transform);
-} else {
-  startClient();
-}
